# openapi.yaml
openapi: 3.1.0
info:
  title: Shoutboard API
  version: 1.0.0
  description: |
    API-first employee recognition group card platform. Shoutboard enables teams to
    create collaborative recognition cards for birthdays, promotions, farewells, and
    other occasions with public shareable links and team contributions.

    ## Key Features
    - **Multi-occasion support**: Birthday, anniversary, promotion, welcome, farewell, custom
    - **Public sharing**: Generate shareable links for public contribution
    - **Team collaboration**: Collect messages from multiple contributors anonymously or identified
    - **Webhook integrations**: Real-time event notifications for board creation, posting, and delivery
    - **API-first design**: Full programmatic access with JWT or API key authentication
    - **Analytics**: Track board engagement, contributor metrics, and team participation

    ## Authentication Methods
    The API supports two authentication methods:

    ### JWT Bearer Token
    Register and login to receive a JWT token valid for 7 days. Include in Authorization header:
    ```
    Authorization: Bearer <jwt_token>
    ```

    ### API Key
    Create persistent API keys in the dashboard for service-to-service integrations.
    Include in header:
    ```
    X-API-Key: sb_live_<random_string>
    ```
    API keys can be scoped to specific permissions and revoked at any time.

    ## Webhook Events
    Subscribe to real-time events:
...

...

    ## Rate Limiting
...

...
    ```json
...

...

...
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login or /auth/register

    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for service-to-service authentication
  schemas:
...
    ApiKey:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        prefix:
          type: string
          description: First 12 chars of key for display (e.g., sb_live_xxxxx)
        scopes:
          type: array
          items:
            type: string
          example: [boards:read, boards:write]
        lastUsedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        key:
          type: string
          description: Full API key (only returned on creation)
      required: [id, name, prefix, scopes, createdAt]
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
          required: [code, message]
      required: [success, error]
...
paths:
...
  /v1/api-keys:
    get:
      summary: List API keys (admin only)
      operationId: listApiKeys
      tags: [API Keys]
      security:
        - bearerAuth: []
      responses:
        200:
          description: List of API keys
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      keys:
                        type: array
                        items:
                          $ref: '#/components/schemas/ApiKey'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        403:
          $ref: '#/components/responses/ForbiddenError'

    post:
      summary: Create API key (admin only)
      operationId: createApiKey
      tags: [API Keys]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 100
                  example: Production Integration
                scopes:
                  type: array
                  items:
                    type: string
                  default: [boards:read, boards:write]
                  example: [boards:read, boards:write, posts:read]
              required: [name]
      responses:
        201:
          description: API key created (full key returned once)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    allOf:
                      - $ref: '#/components/schemas/ApiKey'
                      - type: object
                        properties:
                          key:
                            type: string
                            description: Full API key - STORE THIS SECURELY
                            example: sb_live_abc123def456xyz789
        403:
          $ref: '#/components/responses/ForbiddenError'
  /v1/api-keys/{id}:
    delete:
      summary: Revoke API key (admin only)
      operationId: revokeApiKey
      tags: [API Keys]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: API key revoked
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                      revokedAt:
                        type: string
                        format: date-time
        403:
          $ref: '#/components/responses/ForbiddenError'
        404:
          $ref: '#/components/responses/NotFoundError'

  # Webhook Endpoints
...
  /v1/analytics:
    get:
      summary: Get organization analytics
      operationId: getOrgAnalytics
      tags: [Analytics]
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      description: High-level metrics for the organization.
      responses:
        200:
          description: Organization analytics
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      totalBoards:
                        type: integer
                        example: 42
                      totalPosts:
                        type: integer
                        example: 156
                      totalContributors:
                        type: integer
                        description: Unique identified contributors
                        example: 34
                      totalAnonymousContributions:
                        type: integer
                        example: 18
                      boardsSentThisMonth:
                        type: integer
                        example: 8
                      avgPostsPerBoard:
                        type: number
                        format: float
                        example: 3.71
  /v1/analytics/boards/{id}:
    get:
      summary: Get board-level analytics
      operationId: getBoardAnalytics
      tags: [Analytics]
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      description: Detailed metrics for a specific board.
      responses:
        200:
          description: Board analytics
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      viewCount:
                        type: integer
                      postCount:
                        type: integer
                      uniqueContributors:
                        type: integer
                      anonymousContributions:
                        type: integer
                      postsPerDay:
                        type: object
                        additionalProperties:
                          type: integer
                        description: Posts per day timeline
                        example:
                          '2024-03-20': 5
                          '2024-03-21': 8
                          '2024-03-22': 3
                      boardStatus:
                        type: string
                        enum: [DRAFT, ACTIVE, SENT]
                      createdAt:
                        type: string
                        format: date-time
                      sentAt:
                        type: string
                        format: date-time
                        nullable: true
        404:
          $ref: '#/components/responses/NotFoundError'
components:
  responses:
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: NOT_FOUND
              message: Board not found
    UnauthorizedError:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: UNAUTHORIZED
              message: Authentication required
    ForbiddenError:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: FORBIDDEN
              message: Insufficient permissions
...